backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.1
django-timezone-field==7.1
drf-nested-routers==0.95.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
celery==5.5.3
django-celery-beat==2.8.1
billiard==4.2.2
vine==5.1.0
kombu==5.5.4
redis==5.0.1
channels==4.3.1
channels-redis==4.2.0
daphne==4.0.0
asgiref==3.9.2
requests==2.32.5
urllib3==2.5.0
googlesearch-python==1.2.3
google_search_results==2.4.2
python-decouple==3.8
python-dotenv==1.1.1
python-magic==0.4.27; sys_platform != "win32"
python-magic-bin==0.4.14; sys_platform == "win32"
whitenoise==6.11.0
Pillow==11.1.0
sympy==1.14.0
mpmath==1.3.0
python-crontab==3.3.0
cron_descriptor==2.0.6
python-dateutil==2.9.0.post0
simplejson==3.20.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
wcwidth==0.2.14
six==1.17.0
setuptools>=70
certifi==2025.8.3
idna==3.10
typing_extensions==4.15.0
tzdata==2025.2
packaging==25.0
charset-normalizer==3.4.3
itsdangerous==2.2.0
prompt_toolkit==3.0.52
publicsuffix2==2.20191221
gunicorn==21.2.0
flower==2.0.1
stripe==8.5.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==7.0.0
